requirements.txt
Django==3.2.11
psycopg2==2.9.3
celery==5.2.3
djangorestframework==3.13.1
drf-nested-routers==0.93.4
redis==4.1.2
requests==2.27.1
Unidecode==1.3.2